[bump] report which image/repo when errors happen.
The specific case was an artifactory image that no longer existed. This
caused a 404 error for that image that resulted in the full bump command
failing but there was no way to know which image from the spec was
failing. This intercepts the queries and reports which image and/or repo
has the error.